Added velocity PID controller

+++ control/include/pid_control.h
@@ -7,7 +7,7 @@ class PID_Control {
 
 public:
     PID_Control(float p, float i, float d, int min, int max, float timer_interval = 1, int adjust = 0);
-    float compute(float desired, float actual);
+    int compute(float desired, float actual);
 
 private:
     float kp;
